Extract the signal layer to lib/signals.js
Verbatim move of recordEvent, pruneOldEventFiles, bumpSkillUsage,
normalizeDocsGapTopic, and their retention/stopword constants. The only
conversion: both writers resolve getWorkspace() at use time instead of
reading the root's workspace variable, so signals always land in the
current workspace (pinned by the new seam tests, alongside identity
re-export through _internal). No dependency wiring needed: the module
uses only fs, path, and the config seam; call direction stays root to
lib. appendTranscript and formatTranscript stay in the root: they weave
the signal layer with the store primitives and the search reconcile.
